% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFMfunction.R
\name{RFMfunction}
\alias{RFMfunction}
\title{RFMfunction}
\usage{
RFMfunction(
  data,
  weight_recency = 1,
  weight_frequency = 1,
  weight_monetary = 1
)
}
\arguments{
\item{data}{a data object with columns TransDate, Customer, PurchAmount}

\item{weight_recency}{a specification for weight of recency (default value = 1)}

\item{weight_frequency}{a specification for weight of frequency (default value = 1)}

\item{weight_monetary}{a specification for weight of monetary value (default value = 1)}
}
\value{
The Customers with there groups based on the RFM calculation for a given weights \code{weight_recency} \code{weight_frequency} \code{weight_monetary}
}
\description{
Function that returns the recency, frequency and monetary value KPI.
}
\details{
\code{data} contains the transaction data. The data set must contain a
            column labeled "TransDate" that is configured as dates
            The column "PurchAmount" specifies the total spending per purchase.
            The column "Customer" contains cumstomer-ids
}
\examples{
data(transactions)
RFMfunction(transactions,60,20,20)

}
